+++ src/components/product/descriptions/product-notice.tsx
